sh import last interval as double (for <1h)

+++ RX_FSK/src/Sonde.h
@@ -225,7 +225,7 @@ struct st_sondehub {
         int fiactive;
 	int fiinterval;
 	int fimaxdist;
-	int fimaxage;
+	double fimaxage;
 };
 
 // to be extended
